"Added create, update and delete methods to TecnicoService and implemented creation of new technician in ColaboradoresCreateComponent"
This commit message describes the changes made in the code diff above. It mentions the addition of create, update, and delete methods to the TecnicoService and the implementation of the creation of a new technician in the ColaboradoresCreateComponent. The message is concise and accurately describes the changes made in the code.