Navigation change in NavComponent and addition of edit and delete icons in colaboradores-list.component.html
In NavComponent, the navigation route has been changed from "colaboradores/create" to "colaboradores". In colaboradores-list.component.html, edit and delete icons have been added as actions for each element in the list. The edit icon has a class of "material-icons edit" and the delete icon has a class of "material-icons delete".